requirements.txt
discord.py
asyncio
fivem
youtube-dl
pysqlite3
